Change reason category columns from BIGINT to SMALLINT

+++ crates/database_profile/src/current/write/profile/moderation.rs
@@ -45,7 +45,7 @@ impl CurrentWriteModeration<'_> {
             .do_update()
             .set((
                 state_type.eq(excluded(state_type)),
-                rejected_reason_category.eq(None::<i64>),
+                rejected_reason_category.eq(None::<i16>),
                 rejected_reason_details.eq(None::<String>),
                 moderator_account_id.eq(None::<AccountIdDb>),
                 created_unix_time.eq(excluded(created_unix_time)),
@@ -84,7 +84,7 @@ impl CurrentWriteModeration<'_> {
                 .do_update()
                 .set((
                     state_type.eq(excluded(state_type)),
-                    rejected_reason_category.eq(None::<i64>),
+                    rejected_reason_category.eq(None::<i16>),
                     rejected_reason_details.eq(None::<String>),
                     moderator_account_id.eq(None::<AccountIdDb>),
                     created_unix_time.eq(excluded(created_unix_time)),
